Avoid lazy-loading errors on instance_type
When instances are launched with a source group, the instance_type
is not lazy_loaded in properly.

The issue is intermittent because if another query has already loaded
the instance with a joined instance_type or the instance has not
yet been expunged from the session, then the load succeeds.

Therefore this failure only appears under heavy load. Adding an
explicit joindload fixes the issue.

Change-Id: I450c598c31c416c7be341cbb8cd14e33c6bb2aec