Add call to reset quota usage
There are certain circumstances where we may need to indicate an
uncertainty in a user's current usage of a given resource, thus
forcing a recount of those resources currently in use.  This adds
a QUOTAS.usage_reset() call which can be used to do this.  It
works by resetting the current usage to -1, which will force the
usage record to be refreshed the next time that user reserves
those resources.

Change-Id: I4a40c8edcc78ed9a7d18a293d4d7d6418b04b7f2