More HostAPI() cleanup for cells.
This patch contains further changes needed to support HostAPI() with
cells.  This moves the formatting of hosts API extension responses back
to the extension itself.  Methods for getting service entries have been
added to HostAPI() that can be overriden in a subclass for cells.

Also contains a performance improvement for 'show' in hosts extension.

Change-Id: Ifd8b68ff8d7495f537dff5e01bd3df75c6c5930a