instance_info_cache_update creates wrongly
If instance_info_cache_update() finds that the entry doesn't exist, it
attempts to create an entry, but it didn't use all of the data. This
patch fixes that.

Also if the info_cache entry was found to be deleted, the old code would
attempt to create a new entry.  This would have failed with
DBDuplicateEntry..  Instead, raise InstanceInfoCacheNotFound, which is
more appropriate. (It doesn't seem we actually ever hit this case.)

Fixes bug 1131594

Change-Id: Id008faf5c4a48ab6a04e75796b7e61dea102acc9