Force resource updates to update updated_at
When there's no changes in resources, compute_node_update (and other
DB update calls) won't modify 'updated_at'.

'updated_at' is what is used to invalidate the cache in the scheduler's
host_manager.  Because of a race with the compute manager, the scheduler
could be out of sync with the compute_nodes table but have a newer time
on its cache.

By always updating 'updated_at' on resource updates, the periodic task
will be sure to invalidate any bad cache the scheduler has.

Fixes bug 1153778

Change-Id: I19b51a5b84f472cd0f4de6460a4edb540cc62da2