API extensions framework for v3 API
This is the initial patch for the new extension framework to be used
by the Nova v3 API. It will only be used by v3 API extensions
and not v2 API extensions. v3 API extensions will only use this plugin
framework and will not be compatible with the old one.

- Sets up a /v3 url prefix
- Looks in an entry point namespace of nova.api.extensions
- The fixed_ips extensions is ported to /v3 as an example of
  a resource extensions. Required changes are very minor.
- All extensions must derive from the V3APIExtensionBase class
- Drops updated field from extensions, replaced with version field
- Ports tests for fixed_ips extension
- None of the core has been ported in this patch, though
  the example extension works without it. The intent is
  to port the core code over as plugins as well. There will still
  be a conceptual core however I don't think we need a separate directory
  for core.

This is the first of a series of patches to add support for the
new extension framework. Future direction including support
for controller extensions, removal of extension code in core
code etc can be seen here:

https://github.com/cyeoh/nova/tree/v3_api_extension_framework

Partially implements blueprint v3-api-extension-framework

Change-Id: I88aa6353ad1d74cac51abbb6aac7274b1567485a