Catch ldap ImportError
Since the `ldap` module isn't a requirement, we need to guard against it not
being present.

One reason this is needed is that we generate sample configs by walking the
module import tree. If a module isn't importable--because a depenency isn't
met, for example--the module's configs are omitted from the sample config,
requiring the developer to manually add them back.

Fixes bug 1208560

Change-Id: Ie23b7fad118ecadabf7e36703934fa933f4e8c50