Convert TestCases to NoDBTestCase
NoDBTestCase is faster because it skips initializing a database for
tests which don't actually talk to the database. Its also nicer
because its more explicit about the dependancies of a given test.
This is a second pass at moving test cases over to the database
lacking version, but further work is required here.

Partial-Bug: 1227473

Change-Id: I16313b8febb8a3c681fdf7e7278cc5a09138e542