Refactor return value of fixed_ip_associate calls
This makes the fixed_ip_associate* calls more object-friendly.
Rather few things outside of the tests care about the return value of
these methods, so this is less invasive than it looks. All of the
tests and manager methods that this touches will be replaced
as part of the objectification going forward, so this is only
transitional.

Related to blueprint nova-network-objects

Change-Id: I4764fa0bfca355377d2642ddf3fbfe17cf5b761e