Add new db api get functions for ec2_snapshot
This patch adds two new getters for EC2 snapshot mappings to the db
API.  The existing getters don't return the full db model.  Instead,
they only return an ID or UUID.  The full db model is needed when this
gets converted over to a nova object.

Related to blueprint convert-ec2-api-to-use-nova-objects

Change-Id: I8e2ed5086a7768f373d35c2f34386f2d30d11e07