Add instance_extra_update_by_uuid() to DB API
Since we have multiple things reading from and writing to the
instance_extra table, we really need a way to just update the value
in a single column, by instance_uuid, without looking it up first,
without wiping the other columns, or updating them with stale data
introducing a race.

This adds the new API, updates instance_create() to pre-create the
extra record, and makes the numa_topology object do this sort of
update instead of creating the record itself. As a result of this
change, we no longer need instance_extra_create() exposed out of
the DB API.

Change-Id: I0819cab81467ab41551e8937be7443a7761cbfb8