Revert "Replace outdated oslo-incubator middleware"
This is breaking the Juno -> master grenade path. As we are using the old
api-paste.ini file which requires the oslo-incubator copy of request_id

This breakage didn't happen from Icehouse -> master since the icehouse
api-paste.ini didn't have request_id as a middleware. request_id is only
used as middleware in the V21/V3 API and not the V2 API.

Closes-Bug: #1383757
This reverts commit 641de5666f38532252e91735a1cd2611d96c489a.

Change-Id: I1ca6bddf9d9370529e999efcc7eae60622e19e6f