dummy patch to let tox functional pass
Tox functional job currently will fail because there are no tests,
which is makes it a little hard to atomically turn on voting with the
test move. To trick the system we can make tox functional always pass,
then change the tox definition to the real thing during the sideways
move.

Part of bp:functional-tests-for-nova

Change-Id: Ic276522851fba6d1d95dd1f3bdde1d48935568a3