extract RPC setup into a fixture
This extracts the common RPC setup into a fixture. Eventually we need
to get this fixture off the base setup as < 5% of tests actually use
the RPC infrastructure today (probably more than half of those are
using it by accident due to insufficient mocking), and only 3 cells
tests actually need the exmods setting.

This moves the ball forward slightly towards getting there.

part of bp:functional-tests-for-nova

Change-Id: I7ab9f7f0615c137025aa07e4113e0d207d36a4c5