Fix inefficient transaction usage in floating_ip_bulk_destroy
floating_ip_bulk_destroy was creating a session, then not using it
while it iterated over given ip ranges. Instead, it was creating a new
session for both retrieving and deleting floating ips.

Because these updates now happen in a transaction, they must be
committed before calling the reservations api, which expects to be
able to see the changes.

Note that this function still races:

* between querying floating ips and deleting floating ips
* between populating project_id_to_quota_count and updating quota

It needs to be re-architected. This change simply makes the current
behaviour more efficient.

Change-Id: I25635eeb6c85f834f740d7b4501b2f4df773a258