Remove db layer hard-code permission checks for service_get_all
This patches remove the hard-code permission checks for db call
service_get_all.

There are elevated contexts in the code used to call those methods which
aren't removed in this patch. The elevated context will be cleaned up in
last step. This is safe if we forget removing some hard-code permission.

Partially implements bp v3-api-policy

Change-Id: I4f642d67f2837655189afb2640a062b8e591943d