refactor policy fixtures to allow use of real policy
The policy_fixture previously contained 2 fixtures. ``PolicyFixture``
which dumped fake_policy into a policy.json and loaded it, and
``RoleBasedPolicyFixture`` which uses the real policy file, and
rewrites all the permissions for role based testing. There was no
option to just use the real policy file.

This refactors the fixtures to act as follows:

``RealPolicyFixture`` - use the real policy, provide a hook to
_prepare_policy before we load it, and make policy name an instance
variable that can be overwritten.

Make ``PolicyFixture`` and ``RoleBasedPolicyFixture`` inherit from it,
and put their customizing in _prepare_policy. In both cases they use a
temp file for the policy file that's loaded to ensure we're not
overwriting the default.

Documentation about intent and usage for these classes is also added.

Change-Id: I8bafe7c9949fc1a16da1226886b8835c99e81c40