Fix multipath iSCSI encrypted volume attach failure
Currently iSCSI volume attachment fails if iscsi_use_multipath is
set to True. This is because the encryptor requests cryptsetup
to create the symlink to the LUKS device with the same name of
the device-mapper multipath device. To avoid the name collision,
this patch adds the 'crypt-' prefix to the symlink.

Change-Id: I01a3104025a9c479e40933ca24024355df9dfb8b
Closes-Bug: #1439869