docs: add test strategy and feature classification
This effort is trying to ensure we better document what is currently
tested and know to work, and what is not currently tested.

This renames the Hypervisor support matrix, to the Feature support matrix.
The vision is to move the support matrix ticks to appear only for
features that have tests passing. To enable this to happen, the column
will change from being the virt driver, to being a specific combination
of technologies (such as libvirt + KVM + ceph + neutron ML2 with ovs)

The second step is to include information about the maturity of the
specific feature that is being tested. This will mean the matrix rows
will instead reference a feature group, that has an associated list of
tempest test uuid and links to detailed API docs.

Change-Id: Ia2d489cb4e1fd57737468df4f9fc10e9ad8c011c