use_neutron_default_nets: StrOpt ->BoolOpt
Config option use_neutron_default_nets is StrOpt with
value 'True' or 'False'.  But current method _test_network_index
uses it as boolean type, this lead to can't pass when comparing it
with string type 'True', it doesn't test properly. This commit
makes use_neutron_default_nets as BoolOpt.

DocImpact: This option is now a BoolOpt and documentation should
be updated accordingly.

Change-Id: I19a57db073359a9e58a16cd0de39d39aa95d2aa5