Workaround reno reverts by accepting warnings
b2acc9fa864b6fe10bc0c5f3786b976b472b1b27 on stable/liberty introduced a
sphinx warning which was not treated as an error in stable/liberty nova
releasenotes tox runs so it slipped in, but it is treated as an error
in master branch nova releasenotes tox runs, and since reno scans the
branches, master is broken because of the error in stable/liberty.

Plus, reno scans the git history, so right now it's not obvious how to
travel back in time and fix the stable/liberty change.

This change reduces the level of control we have by not treating reno warnings
as errors. This is not fun but it's the only way we can still land
release notes until an appropriate fix is merged (in reno or nova).

Again, this is a temporary patch and we consider reverting it once bug 1534613
is fixed.

Change-Id: Iedb69c66632383e056840e3294a90cda651a8778
Related-Bug: #1534613