Use db connection from RequestContext during queries
Now that db connections can be specified through the RequestContext,
any queries on tables that reside in cell dbs must use the methods
provided to create transaction context managers on-the-fly using the
db connection from the RequestContext instead of using the
"main_context_manager" for all accesses. The get_context_manager()
method returns the main_context_manager if no db connection is
found in the RequestContext, so things continue to work even when
no cell dbs exist.

The tables that reside in cell dbs are based on the Cells v2
database split doc [1].

Related to blueprint cells-db-connection-switching

[1] https://review.openstack.org/#/c/277543

Change-Id: I4452116ae0c88ba18672839f2856247092bf6b9a